+++ src/api/index.js
@@ -5,13 +5,15 @@ import MatchesAPI from './Matches'
 import MatchmakingAPI from './Matchmaking'
 import NotificationsAPI from './Notifications'
 import ProfilesAPI from './Profiles'
+import SupportAPI from './Support'
 
 export {
-  NotificationsAPI,
   AccountsAPI,
-  MatchmakingAPI,
   FriendsAPI,
-  MatchesAPI,
-  ProfilesAPI,
   LobbiesAPI,
+  MatchesAPI,
+  MatchmakingAPI,
+  NotificationsAPI,
+  ProfilesAPI,
+  SupportAPI,
 }