+++ src/components/account/InactivateAccountCard.jsx
@@ -69,6 +69,7 @@ export default function InactivateAccountCard() {
           isOpen={isOpenModal}
           title="INATIVAR CONTA"
           onClose={handleClose}
+          size="3xl"
         >
           <Container justify="center" align="center" column gap={40}>
             <Text color="gray.700" fontSize={14} textAlign="center">