+++ src/components/support/ConfirmationContent.stories.jsx
@@ -1,0 +1,10 @@
+import { ConfirmationContent } from '@components'
+
+export default {
+  title: 'Support/ConfirmationContent',
+  component: ConfirmationContent,
+}
+
+export const Default = {
+  render: (props) => <ConfirmationContent {...props} />,
+}