+++ src/setupTests.js
@@ -3,6 +3,7 @@
 // expect(element).toHaveTextContent(/react/i)
 // learn more: https://github.com/testing-library/jest-dom
 import '@testing-library/jest-dom'
+const nodeCrypto = require('crypto')
 
 window.matchMedia =
   window.matchMedia ||
@@ -13,3 +14,9 @@ window.matchMedia =
       removeListener: function () {},
     }
   }
+
+window.crypto = {
+  getRandomValues: function (buffer) {
+    return nodeCrypto.randomFillSync(buffer)
+  },
+}